cloud_functions/license_check/requirements.txt
Flask
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2 